requirements.txt
openai>=1.0.0
numpy>=1.24.0
soundcard>=0.4.2
soundfile>=0.12.1
rich>=13.0.0
python-dotenv>=1.0.0